Basic Player Movment(NO ANIMATION)
Created the player controller and added it to the CyberUwu model, added two scripts, player movement, and mouse look, which control both the physical movement of the player along the X and Z axis, as well as the anility to turn the character by moving the mouse. Also added a placeholder splash screen image for the game builds.