Added Animation Dependencies
Imported and installed the Unity Animation Rigging utility that allows us to rig and animate models inside of the editor. Then proceeded to mess around with a super simple "Tracking" that makes the characters head follow a target, in this case, a cube, however encountered plenty of problems, will get back on it later after ive slept!